+++ src/app/pages/account-page/account-page.component.ts
@@ -1,4 +1,4 @@
-import {Component, OnInit} from "@angular/core";
+import {Component} from "@angular/core";
 
 import {AuthService} from "../../services/auth.service";
 
@@ -7,7 +7,7 @@ import {AuthService} from "../../services/auth.service";
   templateUrl: "./account-page.component.html",
   styleUrls: ["./account-page.component.scss"]
 })
-export class AccountPageComponent implements OnInit{
+export class AccountPageComponent{
   constructor(public auth: AuthService){
   }
 }