+++ src/app/services/settings.service.ts
@@ -1,54 +1,75 @@
 import {Injectable} from "@angular/core";
 
-import {CookieService} from "ngx-cookie-service";
-
 @Injectable({
   providedIn: "root"
 })
 export class SettingsService{
+  static readonly prefix = "setting_";
+
+  static readonly serverColumnsDefault = ["players", "address", "owner", "country", "gameStyle", "title"];
+  static readonly playerColumnsDefault = ["callsign", "team", "score", "winsLosses", "tks"];
+
   readonly serverColumns = ["players", "address", "owner", "protocol", "country", "gameStyle", "title"];
   readonly serverColumnNames = ["Players", "Address", "Owner", "Protocol", "Country", "Game Style", "Title"];
 
   readonly playerColumns = ["callsign", "team", "score", "winsLosses", "tks"];
   readonly playerColumnNames = ["Callsign", "Team", "Score", "Wins / Losses", "Team Kills"];
 
-  get displayedServerColumns(): string[]{
-    if(!this.cookieService.check("serverColumns")){
-      return ["players", "address", "owner", "country", "gameStyle", "title"];
+  private getList(key: string, defaults: string[] = []): string[]{
+    const data = localStorage.getItem(`${SettingsService.prefix}${key}`);
+    if(!data){
+      return defaults;
     }
 
-    return JSON.parse(this.cookieService.get("serverColumns"));
+    return JSON.parse(data);
+  }
+
+  private setList(key: string, value: string[], defaults: string[] = []): void{
+    if(JSON.stringify(value) === JSON.stringify(defaults)){
+      return localStorage.removeItem(`${SettingsService.prefix}${key}`);
+    }
+
+    localStorage.setItem(`${SettingsService.prefix}${key}`, JSON.stringify(value));
+  }
+
+  private getBool(key: string): boolean{
+    return localStorage.getItem(`${SettingsService.prefix}${key}`) === "true" ? true : false;
+  }
+
+  private setBool(key: string, value: boolean): void{
+    if(!value){
+      return localStorage.removeItem(`${SettingsService.prefix}${key}`);
+    }
+
+    localStorage.setItem(`${SettingsService.prefix}${key}`, value.toString());
+  }
+
+  get displayedServerColumns(): string[]{
+    return this.getList("serverColumns", SettingsService.serverColumnsDefault);
   }
   get displayedPlayerColumns(): string[]{
-    if(!this.cookieService.check("playerColumns")){
-      return ["callsign", "team", "score", "winsLosses", "tks"];
-    }
-
-    return JSON.parse(this.cookieService.get("playerColumns"));
+    return this.getList("playerColumns", SettingsService.playerColumnsDefault);
   }
 
   get darkMode(): boolean{
-    return this.cookieService.get("darkMode") === "true" ? true : false;;
+    return this.getBool("darkMode");
   }
   set darkMode(value: boolean){
-    this.cookieService.set("darkMode", value.toString());
+    this.setBool("darkMode", value);
   }
 
   get compact(): boolean{
-    return this.cookieService.get("compact") === "true" ? true : false;
+    return this.getBool("compact");
   }
   set compact(value: boolean){
-    this.cookieService.set("compact", value.toString());
+    this.setBool("compact", value);
   }
 
   get gridView(): boolean{
-    return this.cookieService.get("gridView") === "true" ? true : false;
+    return this.getBool("gridView");
   }
   set gridView(value: boolean){
-    this.cookieService.set("gridView", value.toString());
-  }
-
-  constructor(private cookieService: CookieService){
+    this.setBool("gridView", value);
   }
 
   toggleDisplayedServerColumn(column: string): void{
@@ -64,7 +85,7 @@ export class SettingsService{
       columns.splice(this.serverColumns.indexOf(column), 0, column);
     }
 
-    this.cookieService.set("serverColumns", JSON.stringify(columns));
+    this.setList("serverColumns", columns, SettingsService.serverColumnsDefault);
   }
 
   toggleDisplayedPlayerColumn(column: string): void{
@@ -80,6 +101,6 @@ export class SettingsService{
       columns.splice(this.playerColumns.indexOf(column), 0, column);
     }
 
-    this.cookieService.set("playerColumns", JSON.stringify(columns));
+    this.setList("playerColumns", columns, SettingsService.playerColumnsDefault);
   }
 }